lintWarn.patch by Ticker Berkin: "Here are some fixes to warnings I get when I build with options   -XLint:all -Xlint:-serial -Xlint:-path
also an improvement to where city/zip/highway/exitFac ptr sizes are calculated when POIs are written"

+++ src/uk/me/parabola/imgfmt/app/net/RouteArc.java
@@ -317,7 +317,7 @@ public class RouteArc {
 
 		// We write this big endian
 		if(log.isDebugEnabled())
-			log.debug("val is", Integer.toHexString((int)val));
+			log.debug("val is", Integer.toHexString(val));
 		writer.put1u(val >> 8);
 		writer.put1u(val & 0xff);
 	}
